Move decoupling classes to `paymentsheet` module
