+++ link/src/main/java/com/stripe/android/link/injection/LinkCommonModule.kt
@@ -1,7 +1,5 @@
 package com.stripe.android.link.injection
 
-import com.stripe.android.DefaultIntentConfirmationInterceptor
-import com.stripe.android.IntentConfirmationInterceptor
 import com.stripe.android.Stripe
 import com.stripe.android.core.Logger
 import com.stripe.android.core.injection.IOContext
@@ -28,12 +26,6 @@ internal interface LinkCommonModule {
     @Binds
     @Singleton
     fun bindLinkEventsReporter(linkEventsReporter: DefaultLinkEventsReporter): LinkEventsReporter
-
-    @Binds
-    @Singleton
-    fun bindIntentConfirmationInterceptor(
-        impl: DefaultIntentConfirmationInterceptor,
-    ): IntentConfirmationInterceptor
 
     companion object {
         @Provides
